refactor(global): refactor, add go_router for routing and widgetbook generator
Moved to widgetbook 3.x beta which fixed the issue of wrongly scaling.
