+++ lib/core/shared/presentation/theming/constants/border.dart
@@ -1,0 +1,4 @@
+class BorderConstants {
+  static double borderRadius = 12;
+  static double buttonRadius = 8;
+}
