+++ lib/core/shared/presentation/theming/themes/theme.dart
@@ -1,0 +1,2 @@
+export 'dark_theme.dart';
+export 'light_theme.dart';
