+++ lib/home_screen/view/home_screen.dart
@@ -9,7 +9,7 @@
 
 import 'package:flutter/material.dart';
 import 'package:fpb/assets/fpb_icons/fpb_icons_icons.dart';
-import 'package:fpb/core/shared/presentation/theming/colors.dart';
+import 'package:fpb/core/shared/presentation/theming/colors/colors.dart';
 
 class MyHomePage extends StatefulWidget {
   const MyHomePage({super.key});
