+++ lib/onboarding/onboarding.dart
@@ -1,1 +1,2 @@
-
+export 'view/onboarding_screens.dart';
+export 'view/splash_screen.dart';
