+++ lib/onboarding/view/splash_screen.dart
@@ -1,0 +1,30 @@
+// a stateless widget that shows the splash screen with a logo and a text
+import 'package:flutter/material.dart';
+
+class SplashScreen extends StatelessWidget {
+  const SplashScreen({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: Center(
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: const [
+            // temporary logo
+            FlutterLogo(size: 200),
+
+            SizedBox(height: 20),
+            Text(
+              'Welcome to FinFlow',
+              style: TextStyle(
+                fontSize: 20,
+                fontWeight: FontWeight.bold,
+              ),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
