+++ lib/router/go_routes.dart
@@ -1,0 +1,36 @@
+import 'package:flutter/material.dart';
+import 'package:fpb/home_screen/home_screen.dart';
+import 'package:fpb/onboarding/onboarding.dart';
+import 'package:go_router/go_router.dart';
+
+/// The route configuration.
+final GoRouter goRouter = GoRouter(
+  routes: <RouteBase>[
+    GoRoute(
+      path: '/',
+      builder: (BuildContext context, GoRouterState state) {
+        return const SplashScreen();
+      },
+    ),
+    GoRoute(
+      path: '/home',
+      builder: (BuildContext context, GoRouterState state) {
+        return const MyHomePage();
+      },
+      // routes: <RouteBase>[
+      //   GoRoute(
+      //     path: 'details',
+      //     builder: (BuildContext context, GoRouterState state) {
+      //       return const DetailsScreen();
+      //     },
+      //   ),
+      // ],
+    ),
+    GoRoute(
+      path: 'getStarted',
+      builder: (BuildContext context, GoRouterState state) {
+        return const OnboardingPage();
+      },
+    ),
+  ],
+);
