+++ pubspec.yaml
@@ -17,8 +17,10 @@ dependencies:
   flutter_localizations:
     sdk: flutter
   flutter_svg: ^1.1.6
+  go_router: ^5.2.4
   intl: ^0.17.0
-  widgetbook: ^2.4.1
+  widgetbook: ^3.0.0-beta.7
+  widgetbook_annotation:
 
 dev_dependencies:
   bloc_test: ^9.1.0
@@ -26,6 +28,7 @@ dev_dependencies:
     sdk: flutter
   mocktail: ^0.3.0
   very_good_analysis: ^3.1.0
+  widgetbook_generator:
 
 flutter:
   uses-material-design: true
