+++ test/onboarding/view/onboarding_screens_test.dart
@@ -1,0 +1,7 @@
+import 'package:flutter_test/flutter_test.dart';
+
+void main() {
+  testWidgets('onboarding screens ...', (tester) async {
+    // TODO: Implement test
+  });
+}
